moving dtos around
