+++ backend/Bartender.Domain/DTO/Staff/LoginStaffDto.cs
@@ -1,0 +1,8 @@
+﻿
+namespace Bartender.Domain.DTO.Staff;
+
+public class LoginStaffDto
+{
+    public string Username { get; set; }
+    public string Password { get; set; }
+}
