+++ backend/Bartender.Domain/Interfaces/IAuthService.cs
@@ -1,0 +1,9 @@
+﻿
+using Bartender.Domain.DTO.Staff;
+
+namespace Bartender.Domain.Interfaces;
+
+public interface IAuthService
+{
+    Task<ServiceResult<string>> LoginAsync(LoginStaffDto loginDto);
+}
