+++ backend/Bartender.Domain/Interfaces/IStaffService.cs
@@ -1,4 +1,4 @@
-﻿using Bartender.Domain.DTO;
+﻿using Bartender.Domain.DTO.Staff;
 
 namespace Bartender.Domain.Interfaces;
