+++ backend/Bartender.Domain/Services/TableInteractionService.cs
@@ -4,7 +4,6 @@ using Bartender.Data.Enums;
 using Bartender.Domain.DTO;
 using Bartender.Domain.Interfaces;
 using Microsoft.Extensions.Logging;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace Bartender.Domain.Services;
