+++ backend/BartenderTests/StaffServiceTests.cs
@@ -2,7 +2,7 @@
 using Bartender.Data.Enums;
 using Bartender.Data.Models;
 using Bartender.Domain;
-using Bartender.Domain.DTO;
+using Bartender.Domain.DTO.Staff;
 using Bartender.Domain.Interfaces;
 using Bartender.Domain.Services;
 using Microsoft.Extensions.Logging;
